// project/aitoearn-wxplat/project/aitoearn-monorepo/apps/aitoearn-channel/src/libs/facebook/facebook.interfaces.ts
export interface FacebookPostCommentsRequest {
  filter?: 'stream' | 'toplevel'
  order?: 'chronological' | 'reverse_chronological'
  since?: string
  summary?: 'order' | 'total_count' | 'can_comment'
  fields: string
  before?: string
  after?: string
}